% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/d_estimation_fnxs.R
\name{calc.stick.logLn}
\alias{calc.stick.logLn}
\title{Wrapper function so log-likelihood of stickbreaking can be extracted by optimize() function}
\usage{
calc.stick.logLn(geno.matrix, fit.matrix, d.vect, wts = c(2, 1))
}
\arguments{
\item{geno.matrix}{Genotype matrix generated in
\code{\link{generate.geno.matrix}}}

\item{fit.matrix}{Fitness matrix generated in
\code{\link{sim.stick.data}}}

\item{d.vect}{Parameter to be maximized over}

\item{wts}{Vector of weights to weight genotypes by. Used when
\code{\link{generate.geno.weight.matrix}} is called (see that function).
Default is \code{c(2,1)}, meaning weight single-mutation genotypes twice as heavily as others.
Alternatively, vector of weights corresponding to geno.matrix can be provided.}
}
\value{
Vector of log-likelihoods
}
\description{
Wrapper function so log-likelihood of stickbreaking can be extracted by optimize() function
}
\details{
Calls \code{fit.stick.model.given.d} to get log likelihood
}

