% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_selection_functions.R
\name{calculate.posteriors.for.datasets}
\alias{calculate.posteriors.for.datasets}
\title{Calculates posterior probabilities for each row of dataset given model}
\usage{
calculate.posteriors.for.datasets(data, mod)
}
\arguments{
\item{data}{Dataset as generated by \code{\link{sim.data.from.priors.for.mod.selection}}
or \code{\link{sim.data.for.mod.selection}}}

\item{mod}{Multinomial regression model from nnet package fit using the \code{\link{fit.nnet.multinomial.regression}} function.}
}
\value{
dataset with the posterior probabilities of add, mult and stick models bound (using cbind)
to right edge of the supplied dataset.
}
\description{
Calculates posterior probabilities for each row of dataset given model
}

