% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_fnxs.R
\name{create.2panel.model.fit.to.data.plot}
\alias{create.2panel.model.fit.to.data.plot}
\title{Create 2 planel plot of observed vs model fit data}
\usage{
create.2panel.model.fit.to.data.plot(outpath, smry.w.probs, preds,
  print.labels = TRUE, pos.v, off.x, off.y, lims, cex.v = c(1, 1, 1), col.v,
  text.cex = 0.75, xlab = "Observed Fitness",
  ylab = "Model Predicted Fitness", smry.w.probs.2, preds.2,
  print.labels.2 = TRUE, pos.v.2, off.x.2, off.y.2, lims.2, cex.v.2 = c(1,
  1, 1), text.cex.2 = 0.75, xlab.2 = "Observed Fitness",
  ylab.2 = "Model Predicted Fitness", leg.loc = c("bottomright",
  "bottomright"), leg.cex = 1, leg.bty = "n", R2.digits = 3,
  P.digits = 2)
}
\arguments{
\item{outpath}{Path including filename to direct plot}

\item{smry.w.probs}{List with fit.smry statistics and posterior probabilties}

\item{preds}{Data frame with genotype rows (rownames are 0/1 strings), predicted fitness under each model
(column headers "stick", "mult" and "add") and observed fitness (column header "obs")}

\item{print.labels}{Print genotype labels. TRUE/FALSE. Default=TRUE.}

\item{pos.v}{Vector of which direction from to offest lables (pos argument in text function)}

\item{off.x}{Vector of offsets in x dimension}

\item{off.y}{Vector of offsets in y dimension}

\item{lims}{Vector of length 2 specifying plot limits (x and y get same limits)}

\item{cex.v}{Size of points for stickbreaking, multiplicative and additive models}

\item{col.v}{Color of points for stickbreaking, multiplicative and additive models}

\item{text.cex}{Size of the labels. Defaults=1.}

\item{xlab, }{ylab Labels for x and y axes. Default xlab="Observed Fitness", ylab="Model Predicted Fitness".}

\item{smry.w.probs.2}{List with fit.smry statistics and posterior probabilties for second plot.}

\item{preds.2}{Data frame for 2nd plot with genotype rows (rownames are 0/1 strings), predicted fitness under each model
(column headers "stick", "mult" and "add") and observed fitness (column header "obs")}

\item{print.labels.2}{Print genotype labels in 2nd plot. TRUE/FALSE. Default=TRUE.}

\item{pos.v.2}{Vector of which direction from to offest lables (pos argument in text function) in 2nd plot.}

\item{off.x.2}{Vector of offsets in x dimension in 2nd plot.}

\item{off.y.2}{Vector of offsets in y dimension in 2nd plot.}

\item{lims.2}{Vector of length 2 specifying plot limits (x and y get same limits) in 2nd plot.}

\item{cex.v.2}{Size of points for stickbreaking, multiplicative and additive models in 2nd plot.}

\item{text.cex.2}{Size of the labels ijn 2nd plot. Defaults=1.}

\item{xlab.2, }{ylab.2 Labels for x and y axes in 2nd plot. Default xlab="Observed Fitness", ylab="Model Predicted Fitness".}

\item{leg.loc}{Location of each of the two legends. Default is "bottomright" for both.}

\item{leg.cex}{Text size in legend. Default = 1.}

\item{leg.bty}{bty argument for the legenes. Default = "n".}

\item{R2.digits}{Number of digits to round R2 values to when printing in legend. Default=3.}

\item{P.digits}{Number of digits to round P scores to when printing in legend. Default=2.}
}
\value{
Nothing. Creates plot.
}
\description{
Create 2 planel plot of observed vs model fit data
}

