% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_fnxs.R
\name{create.coe.estimation.figure}
\alias{create.coe.estimation.figure}
\title{Create figure comparing coefficient estimates under three models}
\usage{
create.coe.estimation.figure(outpath, error.table.list, muts.to.plot,
  coes.to.plot, sigs.to.plot)
}
\arguments{
\item{outpath}{The pathway (including file name) where to write figure to}

\item{error.table.list}{List of error tables from \code{analyze.stick.data.batch} and \code{analyze.mult.add.data.batch}
with names "stick", "mult" and "add" (in taht order).}

\item{muts.to.plot}{Simulated mutation values to plot.}

\item{coes.to.plot}{Simulated coefficient values to plot.}

\item{sigs.to.plot}{Simulated sigma values to plot.}
}
\description{
Create figure comparing coefficient estimates under three models
}

