% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_fnxs.R
\name{create.d.estimation.figure}
\alias{create.d.estimation.figure}
\title{Create multipanel figure that compares d esimators}
\usage{
create.d.estimation.figure(outpath, error.table, seq.method.props, mut.vals,
  coe.vals, sig.vals, ylims, est.to.plot, col.v, angle.v, den.v)
}
\arguments{
\item{outpath}{The pathway (including file name) where to write figure to}

\item{error.table}{Generated in \code{analyze.stick.data.batch}.}

\item{seq.method.props}{Generated in \code{analyze.stick.data.batch}}

\item{ylims}{Vector of upper y-axis limits for [1] Failure rate, [2] Bias and [3] rMSE}

\item{est.to.plot}{Vector of estimators to plot. Options: "MLE", "RDB.all", "RDB", "max", "seq" (case sensitive).}

\item{col.v}{Color vector corresponding to est.to.plot}

\item{angle.v}{Angle vector corresponding to est.to.plot}

\item{den.v}{Density vector corresponding to est.to.plot}

\item{muts.to.plot}{Mutation values to plot (values not indecies)}

\item{coes.to.plot}{Coefficient values to plot}

\item{sigs.to.plot}{Sigma values to plot}
}
\value{
Nothing
}
\description{
Create multipanel figure that compares d esimators
}
\details{
Generates figure from paper. This probably will not be included in the package. But I'm including it
here to keep everything together. Creates a pdf and a svg version.
}

