% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_fnxs.R
\name{create.model.fit.to.data.plot}
\alias{create.model.fit.to.data.plot}
\title{Create plot of observed vs model fit data}
\usage{
create.model.fit.to.data.plot(outpath, fit.smry, preds, print.labels = TRUE,
  pos.v, off.x, off.y, lims, cex.v = c(1, 1, 1), col.v, text.cex = 0.75,
  xlab = "Observed Fitness", ylab = "Model Predicted Fitness")
}
\arguments{
\item{outpath}{Path including filename to direct plot}

\item{fit.smry}{Summary of fit statistics from}

\item{preds}{Data frame with genotype rows (rownames are 0/1 strings), predicted fitness under each model
(column headers "stick", "mult" and "add") and observed fitness (column header "obs")}

\item{print.labels}{Print genotype labels. TRUE/FALSE. Default=TRUE.}

\item{pos.v}{Vector of which direction from to offest lables (pos argument in text function)}

\item{off.x}{Vector of offsets in x dimension}

\item{off.y}{Vector of offsets in y dimension}

\item{lims}{Vector of length 2 specifying plot limits (x and y get same limits)}

\item{cex.v}{Size of points for stickbreaking, multiplicative and additive models}

\item{col.v}{Color of points for stickbreaking, multiplicative and additive models}

\item{text.cex}{Size of the labels. Defaults=1.}

\item{xlab, }{ylab Labels for x and y axes. Default xlab="Observed Fitness", ylab="Model Predicted Fitness".}
}
\value{
Nothing. Creates plot.
}
\description{
Create plot of observed vs model fit data
}

