% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_fnxs.R
\name{create.posterior.wireframe.individual}
\alias{create.posterior.wireframe.individual}
\title{Create individual wireframe figures of posterior probabilities from analyzed simulated data}
\usage{
create.posterior.wireframe.individual(posterior.list, outpath,
  coes.to.plot = NA, sigs.to.plot = NA, plot.type = "mean.post")
}
\arguments{
\item{posterior.list}{List of posterior probability dataframes generated in
\code{\link{summarize.posteriors.on.simulated.dataset}}}

\item{outpath}{File (with full path) where plot is written}

\item{coes.to.plot}{Coefficients to plot. If \code{NA} it plots all values in posterior.list.}

\item{sigs.to.plot}{Sigma values to plot. If \code{NA} it plots all values in posterior.list.}

\item{plot.type}{Specifies which field of posterior.list to plot on z-axis of wireframe.
Options are "mean.post" (mean posterior probabiltiy) or "true.unq" (proportion of time model is only accepted model).
Default is "mean.post".}
}
\value{
Nothing. Creates pair of plots (pdf and svg) for each model/mutation
combination written to \code{outpath}
}
\description{
Create individual wireframe figures of posterior probabilities from analyzed simulated data
}

