% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_fnxs.R
\name{create.stick.coe.estimation.figure}
\alias{create.stick.coe.estimation.figure}
\title{Create multipanel figure showing bias and rMSE when estimating stickbreaking coefficients}
\usage{
create.stick.coe.estimation.figure(outpath, error.table, muts.to.plot,
  coes.to.plot, sigs.to.plot)
}
\arguments{
\item{outpath}{The pathway (including file name) where to write figure to}

\item{error.table}{Data frame with rMSE, bias and other summary statistics. Generated in \code{analyze.stick.data.batch}.}

\item{muts.to.plot}{Simulated mutation values to plot.}

\item{coes.to.plot}{Simulated coefficient values to plot.}

\item{sigs.to.plot}{Simulated sigma values to plot.}

\item{seq.method.props}{Generated in \code{analyze.stick.data.batch}}
}
\value{
Nothing
}
\description{
Create multipanel figure showing bias and rMSE when estimating stickbreaking coefficients
}
\details{
Generates figure for paper. This probably will not be included in the package. But I'm including it
here to keep everything together. Creates a pdf and a svg version.
}

