% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_fnxs.R
\name{empty.plot}
\alias{empty.plot}
\title{Create a blank plot to add to}
\usage{
empty.plot(mars = c(0, 0, 0, 0))
}
\arguments{
\item{mars}{Margins to set in \code{par(mar=mars)}. Default is c(0,0,0,0).}
}
\value{
Nothing
}
\description{
Create a blank plot to add to
}
\details{
Creates an blank plot with no margins and xlim=c(0,1) and ylim=c(0,1).
}

