% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_data_fnxs.R
\name{generate.geno.matrix}
\alias{generate.geno.matrix}
\title{Generate genotype matrix for given number of mutations.}
\usage{
generate.geno.matrix(n)
}
\arguments{
\item{n}{Number of mutations.}
}
\value{
Matrix with all genotypes made of \code{n} mutations. Each column is a mutation (0=absent, 1=present), rows are genotypes.
}
\description{
Generate genotype matrix for given number of mutations.
}
\details{
Used when simulating data for a complete network. Generally called internally during simulations to create matrix for which fitness values will then be simulated.
Minimum value of \code{n} is 2.
}
\examples{
generate.geno.matrix(5)
}
\seealso{
\code{\link{sim.stick.data}}
}

