% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fnxs_general.R
\name{Khan.data}
\alias{Khan.data}
\title{A dataset containing the fitness values for recombinant Escherichia coli bacteria.}
\format{A list of 6 elements containing integers (1 or 0) that indicate the absence or presence of mutated alleles at 5 loci and a numerical element with relative fitness values.}
\usage{
data(Khan.data)
}
\description{
A dataset containing the fitness values for recombinant Escherichia coli bacteria.
}
\references{
Khan, A. I., D. M. Dinh, D. Schneider, R. E. Lenski and T. F. Cooper, 2011 Negative epistasis between beneficial mutations in an evolving bacterial population. Science 332, 1193-1196.
}

