% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_fnxs.R
\name{plot.circle.color.sector}
\alias{plot.circle.color.sector}
\title{Plot a sector of a cirle}
\usage{
\method{plot}{circle.color.sector}(a, b, r, p, col, den = NA, angle = NA,
  ydistortion = 1, zero.angle = 0, lwd = 1, n.pts = 100)
}
\arguments{
\item{a}{X offset}

\item{b}{Y offset}

\item{r}{Radius}

\item{p}{Proportion of circle to fill}

\item{col}{Fill color}

\item{den}{Desity of fill (it hatching). Default is NA.}

\item{angle}{Angle of hatching (if hatching). Default is NA.}

\item{ydistortion}{Distortion to make circles not streched by plotting}

\item{zero.angle}{Angle to offset filled sector}

\item{lwd}{Line width. Default is 1.}

\item{n.pts}{Number of points to discretize circle into. Default=100.}
}
\value{
Nothing
}
\description{
Plot a sector of a cirle
}
\details{
Creates an blank plot with no margins and xlim=c(0,1) and ylim=c(0,1).
}

