% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_selection_functions.R
\name{simulate.data.calculate.posteriors}
\alias{simulate.data.calculate.posteriors}
\title{Simulate data from priors then use to calcualte posterior probability of models given data}
\usage{
\method{simulate}{data.calculate.posteriors}(fit.smry, data, analysis.name,
  coes.prior, sig.prior, d.range, d.adj.max, wts, n.samps.per.mod, min.R2)
}
\arguments{
\item{fit.smry}{X}

\item{data}{X}

\item{analysis.name}{X}

\item{coes.prior}{X}

\item{sig.prior}{X}

\item{d.range}{X}

\item{d.adj.max}{X}

\item{wts}{X}

\item{n.samps.per.mod}{X}

\item{min.R2}{X}
}
\value{
List:\cr
\code{$posteriors} Posterior probability of each epistasis model. \cr
\code{$multinomial.model} Multinomial model fit to the simulated data used to assign posterior probabilities.
}
\description{
Simulate data from priors then use to calcualte posterior probability of models given data
}

