% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_fitting_functions.R
\name{summarize.fits.for.posterior.calc}
\alias{summarize.fits.for.posterior.calc}
\title{Extracts summary statistics from each model needed for posterior calculation}
\usage{
summarize.fits.for.posterior.calc(fit.stick, fit.mult, fit.add)
}
\arguments{
\item{fit.stick}{List returned from fitting data to stickbreaking model in \code{fit.stick.model.given.d}}

\item{fit.mult}{List returned from fitting data to multiplicative model in \code{fit.mult.model}}

\item{fit.add}{List returned from fitting data to additive model in \code{fit.add.model}}
}
\value{
Vector of summary statistics under each model
}
\description{
Extracts summary statistics from each model needed for posterior calculation
}
\details{
Extracts R2, sigma, log-likelihood and P-statistic under stickbreaking, multiplicative and additive models
}

