% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fnxs_general.R
\name{burns.data}
\alias{burns.data}
\title{A dataset containing the fitness values for recombinant poliovirus viruses.}
\format{A list of 5 elements containing integers (1 or 0) that indicate the absence or presence of mutated alleles at 4 loci and a numerical element with relative fitness values.}
\usage{
data(burns.data)
}
\description{
A dataset containing the fitness values for recombinant poliovirus viruses.
}
\references{
Burns, C.C., Shaw, J., Campagnoli, R., Jorba, J., Vincent, A., Quay, J., and Kew, O. (2006). Modulation of Poliovirus Replicative Fitness in HeLa Cells by Deoptimization of Synonymous Codon Usage in the Capsid Region. J Virol 80, 3259-3272.
}

