% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fnxs_general.R
\name{chou.data}
\alias{chou.data}
\title{A dataset containing the fitness values for recombinant strains for Methylobacterium extorquens
Chou, H., H. Chiu, N. Delaney, D. Segrè and C. Marx, 2011 Diminishing returns epistasis among beneficial mutations decelerates adaptation. Science 332: 1190–1192.}
\format{A list of 5 elements containing integers (1 or 0) that indicate the absence or presence of mutated alleles at 4 loci and a numerical element with relative fitness values.}
\description{
A dataset containing the fitness values for recombinant strains for Methylobacterium extorquens
Chou, H., H. Chiu, N. Delaney, D. Segrè and C. Marx, 2011 Diminishing returns epistasis among beneficial mutations decelerates adaptation. Science 332: 1190–1192.
}

