% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_selection_functions.R
\name{fit.nnet.multinomial.regression}
\alias{fit.nnet.multinomial.regression}
\title{Fit training data to multinomial regression using \code{nnet} package}
\usage{
fit.nnet.multinomial.regression(data, mod.formula)
}
\description{
@param data Dataset to fit
 @export
}

